Refactored code for debug connections
Previously, a chitcpd client would connect to the chitcpd
UNIX socket and, if the first command was anything other
than a DEBUG message, then the connection was assumed to be
a command connection. If a DEBUG message was received, then
it was considered to be a debug connection (only to send
breakpoint notifications back to the client).

This check was done *after* the handler thread was created,
which meant that, if a DEBUG message was received, the thread
was killed since it was unnecessary: we will only be using
the connection to send back messages, not to receive commands.

This made the code unnecessarily convoluted. Now, there is
a new INIT command that must be sent at the beginning of all
connections, specifying whether it is a command connection
or a debug connection (and, if the latter, with the information
necessary to set up the debug monitor)